# docs/swagger.yaml
definitions:
...
  dto.WithdrawalDTO:
    properties:
      amount:
        type: number
      coin_type:
        type: string
      wallet_address:
        type: string
    required:
    - amount
    - coin_type
    - wallet_address
    type: object
  response.ApiResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is my Go Inv API.
  title: My Inv API
  version: "1.0"
paths:
...
  /admin/walletaddress:
    get:
      consumes:
      - application/json
      description: Get all admin wallets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get all admin wallets
      tags:
      - Admin Dashboard
    post:
      consumes:
      - multipart/form-data
      description: Add wallet as admin
      parameters:
      - description: Coinlogo Image
        in: formData
        name: coin_logo
        required: true
        type: file
      - description: QR code  Image
        in: formData
        name: qr_code
        required: true
        type: file
      - description: Coin name
        in: formData
        name: coin
        required: true
        type: string
      - description: Wallet address
        in: formData
        name: address
        required: true
        type: string
      - description: Coin network
        in: formData
        name: network
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Add wallet
      tags:
      - Admin Action
  /admin/walletaddress/{id}:
    delete:
      consumes:
      - application/json
      description: delete selected wallet
      parameters:
      - description: wallet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: delete wallet
      tags:
      - Admin Action
  /admin/withdrawal:
    get:
      description: Fetch all users withdrawals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch withdrawals
      tags:
      - Admin Dashboard
  /admin/withdrawal/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve user withdrawal
      parameters:
      - description: withdrawal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Approve withdrawal
      tags:
      - Admin Action
  /admin/withdrawal/{id}/decline:
    put:
      consumes:
      - application/json
      description: Decline user withdrawal
      parameters:
      - description: withdrawal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Decline withdrawal
      tags:
      - Admin Action
  /admin/withdrawal/pending:
    get:
      description: Fetch all pending user withdrawals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch pending withdrawals
      tags:
      - Admin Dashboard
...
  /user/withdrawal:
    post:
      consumes:
      - application/json
      description: Create a new withdrawal for user
      parameters:
      - description: withdrawal data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.WithdrawalDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: New withdrawal
      tags:
      - User Action
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"